… Cargo.lock
… Cargo.toml
… common/Cargo.toml
… modules/fetch/Cargo.toml
… modules/ingestor/Cargo.toml
… modules/ingestor/src/graph/package/qualified_package.rs
… modules/ingestor/src/graph/sbom/mod.rs
… modules/ingestor/src/graph/sbom/spdx/mod.rs
… modules/ingestor/src/graph/sbom/spdx/tests/mod.rs
… modules/ingestor/src/graph/sbom/spdx/tests/perf.rs
… modules/ingestor/src/graph/sbom/tests.rs
… modules/ingestor/src/service/advisory/csaf/loader.rs
… modules/ingestor/src/service/advisory/osv/loader.rs