Baseline:

```
[INFO  trustify_module_ingestor::graph::sbom::spdx::tests::perf] parse: 2s 641ms 659us 474ns
[INFO  trustify_module_ingestor::graph::sbom::spdx::tests::perf] ingest 1: 2ms 413us 634ns
[INFO  trustify_module_ingestor::graph::sbom::spdx::tests::perf] ingest 2: 14s 688ms 183us 212ns
[INFO  trustify_module_ingestor::graph::sbom::spdx::tests::perf] commit: 390us 927ns
[INFO  trustify_module_ingestor::graph::sbom::spdx::tests::perf] query: 4ms 109us 506ns
test graph::sbom::spdx::tests::perf::ingest_spdx_medium ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 43 filtered out; finished in 18.87s
```

This PR:

```
2024-04-26T14:07:32.672525Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: parse: 2s 492ms 106us 342ns    
2024-04-26T14:07:32.672536Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 1: 2ms 564us 988ns    
2024-04-26T14:07:32.672546Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: ingest 2: 12s 226ms 606us 90ns    
2024-04-26T14:07:32.672555Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: commit: 385us 427ns    
2024-04-26T14:07:32.672569Z  INFO trustify_module_ingestor::graph::sbom::spdx::tests::perf: query: 4ms 289us 159ns    
test graph::sbom::spdx::tests::perf::ingest_spdx_medium ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 43 filtered out; finished in 16.23s
```

The main changes are:

* less cloning
* less nested "get or ingest" calls